Add UI to shaderbg-web
Created the UI (as removed in 'WebGPU working' commit).  Scene properties can now be modified in the web demo.

Added a new UI for the web demo, with information about the program, scene, and a link to the GitHub.

The scene UI no longer sends app events.  Instead, it batches changes into a vector for the calling code to deal with.

Fixed canvas size on load.