Implement settings
Settings like UI theme, tray config and startup parameters now work.  Scene and preference directory are not implemented yet.

Config is now saved after changing it in the settings UI.

Moved the timer into a separate thread, which also handles quitting the app when all windows (background, window and tray) are not running.